# [DI] Optimize Container::get() for perf