@dkarlovi 2017-12-13
This is just an idea that came up reading this PR, but would doing [this](https://veewee.github.io/blog/optimizing-php-performance-by-fq-function-calls/) for the dumped code be a similar easy fix?