@Koc 2017-12-13
Oh, why PHP's opcode does not inline constants values? Is it possible to do something here from PHP side? // @nikic