@javiereguiluz 2017-12-13
@dkarlovi I think we already do that and not only for the container but for our own code too. Example: https://github.com/symfony/symfony/pull/24866