@stof 2017-12-13
@Koc because the constant is defined in a separate file, and OPCache optimizations are per-file (to avoid having the optimized state depending on the list of previously loaded files)