@zilionis 2017-12-13
One stupid question. Why you chose 10 000 000 times, but not 1 000 000 000 (or whatever). So it can be 1000% or 100000% faster. For me this micro optimization is useless. We just lose code quality here.