@nicolas-grekas 2017-12-13
@zilionis really? Can you run the script locally with these different numbers and report back?