@javiereguiluz 2017-12-13
@zilionis I think there's some error in those numbers. If Nico's benchmarks iterate 10 million times and goes from 2s to 1s ... if we increase iterations 10x to 100 million times, I expect time to go from 20s to 10s ... so the gain remains at 100% and it doesn't increase.

In any case, we can of course be wrong, so please share your script so we can check the numebrs again. Thanks!