@nicolas-grekas 2017-12-13
@zilionis and for the second part of your comment, as replied on Twitter (https://twitter.com/nicolasgrekas/status/940885273587707904)

> yes, it matters when this code is run by a large community that expect the framework to be as seamless as possible even in tight loops.