@zilionis 2017-12-13
@nicolas-grekas @javiereguiluz ok, I am feeling stupid now. But anyway but about this micro change 100% looked to big for me. So i made benchmark (same test script). And I got just 15-25% performance increase.