@javiereguiluz 2017-12-13
@zilionis I was curious so I run the benchmarks myself. These are the numbers:

```
BEFORE 3.4896380901337
AFTER  1.6525621414185
```

So in my case it's consistent with 100% performance increase.