@mvrhov 2017-12-13
Sorry, but I find this stupid, now we have a freaking magic number.