@nicolas-grekas 2017-12-13
(PHP 7.2 btw here, didn't try on anything else recently)