@zilionis 2017-12-13
Currently running on work computer (7.0)
before 9.3259789943695
after 8.2541401386261
12.99%