@javiereguiluz 2017-12-13
@mvrhov *"stupid"* and *"freaking"* words are a bit strong. Please let's keep discussions friendly :)

And yes, you are right that this goes against the recommended practices but:

* It provides a huge performance improvement
* We add a useful comment next to the "magic number", so it's impossible to mistake