@stof 2017-12-13
and we do it only for the code belonging to the hot path of the container, not for all places using the constants (all code dealing with compile-time logic has no reason to inline the value)