@dkarlovi 2017-12-13
> because the constant is defined in a separate file, and OPCache optimizations are per-file 

It seems like this specific change might benefit more complex codebases greatly, it makes sense to try and resolve the constant even if the file containing it is not yet loaded.