@zilionis 2017-12-13
7.2
➜  my_project php test.php 
1.7662029266357
0.92619895935059
90.69%