@Koc 2017-12-13
@stof thank you for explanation. No way to copy this constants to class, something like:

```php
const EXCEPTION_ON_INVALID_REFERENCE = ContinerInterface::EXCEPTION_ON_INVALID_REFERENCE;
```
?