@ostrolucky 2017-12-13
Is this really hot path for Symfony 4+? There isn't lot of public services nowadays.