@stof 2017-12-13
@Koc the fact that you reference the constant from the interface has the same issue: OPCache could maybe inline the reference to the constant defined in the current file, replacing it by a direct reference to `ContainerInterface::EXCEPTION_ON_INVALID_REFERENCE` (i.e. our old code). but it still does not have the *value* of `ContainerInterface::EXCEPTION_ON_INVALID_REFERENCE` in this file (as ContainerInterface is in a different file)