@curry684 2017-12-14
Well, given that we already compile the container into an non-humanly readable mess in a generated namespace, why wouldn't we inline all of it? The compiler can look up the final value of those constants, so we can completely inline all of it. Ie.
```php
class appLocalDebugProjectContainer extends Container
{
    const EXCEPTION_ON_INVALID_REFERENCE = 1;
    const NULL_ON_INVALID_REFERENCE = 2;
    const IGNORE_ON_INVALID_REFERENCE = 3;
    const IGNORE_ON_UNINITIALIZED_REFERENCE = 4;
   ...
```