@nicolas-grekas 2017-12-14
Everyone is comming with better ideas, but please run them before, the script is so simple.
Here is what your suggestions triggers:
 > Fatal error: Cannot inherit previously-inherited or override constant EXCEPTION_ON_INVALID_REFERENCE from interface Symfony\Component\DependencyInjection\ResettableContainerInterface in src/Symfony/Component/DependencyInjection/Container.php on line 40