@kelunik 2018-04-05
Constants from the same class should already be inlined by PHP: https://3v4l.org/jM5BA/vld#output.