@nicolas-grekas 2018-04-05
@kelunik not when taking inheritance into account (and the parent class is in another file.)