@stof 2018-04-05
@kelunik these constants are not from the same class but from a parent