@emirb 2018-05-03
Class constants are now inlined at runtime. 
Implemented in master: https://github.com/php/php-src/commit/1a63fa6ec9b0bacbb726e60c3c212e7d97b518c6